requirements.txt
# Core dependencies
fastapi==0.110.0
uvicorn==0.27.1
python-dotenv==1.0.1
pydantic==2.6.1

# OpenAI API
openai==1.12.0

# HTTP client for API requests
httpx==0.26.0
requests==2.31.0

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Optional UI
streamlit==1.31.0
# chainlit==0.7.700  # Uncomment if using Chainlit for UI 